fix: extracted team member invite to a component
